SimpleScribeVA_Offline/requirements.txt
flask
python-dotenv
sounddevice
scipy
openai>=1.0.0
psutil
